+++ sdk/defendereasm/arm-defendereasm/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History
     
+## 1.0.0-beta.2 (2024-08-05)
+Compared with version 1.0.0-beta.1
+    
+**Breaking Changes**
+
+  - Removed operation Workspaces.listBySubscription
+    
+    
 ## 1.0.0-beta.1 (2023-08-16)
 
-The package of @azure/arm-defendereasm is using our next generation design principles. To learn more, please refer to our documentation [Quick Start](https://aka.ms/azsdk/js/mgmt/quickstart ).
+The package of @azure/arm-defendereasm is using our next generation design principles. To learn more, please refer to our documentation [Quick Start](https://aka.ms/azsdk/js/mgmt/quickstart).